create typesafe id types for wave, table and table index

+++ source/xtLib/xtMidiTypes.h
@@ -188,11 +188,11 @@ namespace xt
 
 	namespace Wave
 	{
-		static constexpr uint32_t g_romWaveCount = 300;
-		static constexpr uint32_t g_ramWaveCount = 250;
-		static constexpr uint32_t g_firstRamWaveIndex = 1000;
-		static constexpr uint32_t g_tableCount = 128;
-		static constexpr uint32_t g_wavesPerTable = 64;
+		static constexpr uint16_t g_romWaveCount = 300;
+		static constexpr uint16_t g_ramWaveCount = 250;
+		static constexpr uint16_t g_firstRamWaveIndex = 1000;
+		static constexpr uint16_t g_tableCount = 128;
+		static constexpr uint16_t g_wavesPerTable = 64;
 
 		// these are either algorithmic or invalid, we cannot request them via MIDI
 		static constexpr uint32_t g_algorithmicWavetables[] = {28, 29,
